requirements.txt
pandas>=1.2.1
requests>=2.25.1
python-dotenv>=0.15.0
tqdm>=4.56.0
web3>=5.15.0
dotmap>=1.3.23
